Move using constructors snippets
